Move ORCA code into participants
